+++ price-feeder/oracle/convert.go
@@ -5,6 +5,7 @@ import (
 	"strings"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/rs/zerolog"
 	"github.com/umee-network/umee/price-feeder/config"
 	"github.com/umee-network/umee/price-feeder/oracle/provider"
 	"github.com/umee-network/umee/price-feeder/oracle/types"
@@ -30,10 +31,15 @@ func getUSDBasedProviders(asset string, providerPairs map[string][]types.Currenc
 }
 
 // ConvertCandlesToUSD converts any candles which are not quoted in USD
-// to USD by other price feeds.
+// to USD by other price feeds. It will also filter out any candles not
+// within the deviation threshold set by the config.
+//
+// Ref: https://github.com/umee-network/umee/blob/4348c3e433df8c37dd98a690e96fc275de609bc1/price-feeder/oracle/filter.go#L41
 func convertCandlesToUSD(
+	logger zerolog.Logger,
 	candles provider.AggregatedProviderCandles,
 	providerPairs map[string][]types.CurrencyPair,
+	deviationThresholds map[string]sdk.Dec,
 ) (provider.AggregatedProviderCandles, error) {
 	if len(candles) == 0 {
 		return candles, nil
@@ -71,7 +77,17 @@ func convertCandlesToUSD(
 				if len(validCandleList) == 0 {
 					return nil, fmt.Errorf("there are no valid conversion rates for %s", pair.Quote)
 				}
-				tvwap, err := ComputeTVWAP(validCandleList)
+
+				filteredCandles, err := FilterCandleDeviations(
+					logger,
+					validCandleList,
+					deviationThresholds,
+				)
+				if err != nil {
+					return nil, err
+				}
+
+				tvwap, err := ComputeTVWAP(filteredCandles)
 				if err != nil {
 					return nil, err
 				}
@@ -99,10 +115,15 @@ func convertCandlesToUSD(
 }
 
 // convertTickersToUSD converts any tickers which are not quoted in USD to USD,
-// using the conversion rates of other tickers.
+// using the conversion rates of other tickers. It will also filter out any tickers
+// not within the deviation threshold set by the config.
+//
+// Ref: https://github.com/umee-network/umee/blob/4348c3e433df8c37dd98a690e96fc275de609bc1/price-feeder/oracle/filter.go#L41
 func convertTickersToUSD(
+	logger zerolog.Logger,
 	tickers provider.AggregatedProviderPrices,
 	providerPairs map[string][]types.CurrencyPair,
+	deviationThresholds map[string]sdk.Dec,
 ) (provider.AggregatedProviderPrices, error) {
 	if len(tickers) == 0 {
 		return tickers, nil
@@ -141,7 +162,17 @@ func convertTickersToUSD(
 				if len(validTickerList) == 0 {
 					return nil, fmt.Errorf("there are no valid conversion rates for %s", pair.Quote)
 				}
-				vwap, err := ComputeVWAP(validTickerList)
+
+				filteredTickers, err := FilterTickerDeviations(
+					logger,
+					validTickerList,
+					deviationThresholds,
+				)
+				if err != nil {
+					return nil, err
+				}
+
+				vwap, err := ComputeVWAP(filteredTickers)
 				if err != nil {
 					return nil, err
 				}