+++ x/leverage/client/tests/tests.go
@@ -65,8 +65,8 @@ func (s *IntegrationTestSuite) TestInvalidQueries() {
 			nil,
 		},
 		testQuery{
-			"query lend APY - invalid denom",
-			cli.GetCmdQueryLendAPY(),
+			"query supply APY - invalid denom",
+			cli.GetCmdQuerySupplyAPY(),
 			[]string{
 				"abcd",
 			},
@@ -85,8 +85,8 @@ func (s *IntegrationTestSuite) TestInvalidQueries() {
 			nil,
 		},
 		testQuery{
-			"query loaned - invalid address",
-			cli.GetCmdQueryLoaned(),
+			"query supplied - invalid address",
+			cli.GetCmdQuerySupplied(),
 			[]string{
 				"xyz",
 			},
@@ -95,8 +95,8 @@ func (s *IntegrationTestSuite) TestInvalidQueries() {
 			nil,
 		},
 		testQuery{
-			"query loaned - invalid denom",
-			cli.GetCmdQueryLoaned(),
+			"query supplied - invalid denom",
+			cli.GetCmdQuerySupplied(),
 			[]string{
 				val.Address.String(),
 				fmt.Sprintf("--%s=abcd", cli.FlagDenom),
@@ -148,8 +148,8 @@ func (s *IntegrationTestSuite) TestInvalidQueries() {
 			nil,
 		},
 		testQuery{
-			"query loaned value - invalid address",
-			cli.GetCmdQueryLoanedValue(),
+			"query supplied value - invalid address",
+			cli.GetCmdQuerySuppliedValue(),
 			[]string{
 				"xyz",
 			},
@@ -158,8 +158,8 @@ func (s *IntegrationTestSuite) TestInvalidQueries() {
 			nil,
 		},
 		testQuery{
-			"query loaned value - invalid denom",
-			cli.GetCmdQueryLoanedValue(),
+			"query supplied value - invalid denom",
+			cli.GetCmdQuerySuppliedValue(),
 			[]string{
 				val.Address.String(),
 				fmt.Sprintf("--%s=abcd", cli.FlagDenom),
@@ -271,7 +271,7 @@ func (s *IntegrationTestSuite) TestLeverageScenario() {
 						MaxBorrowRate:        sdk.MustNewDecFromStr("1.5"),
 						KinkUtilization:      sdk.MustNewDecFromStr("0.2"),
 						LiquidationIncentive: sdk.MustNewDecFromStr("0.18"),
-						EnableMsgLend:        true,
+						EnableMsgSupply:      true,
 						EnableMsgBorrow:      true,
 						Blacklist:            false,
 					},
@@ -303,16 +303,16 @@ func (s *IntegrationTestSuite) TestLeverageScenario() {
 			},
 		},
 		testQuery{
-			"query lend APY",
-			cli.GetCmdQueryLendAPY(),
+			"query supply APY",
+			cli.GetCmdQuerySupplyAPY(),
 			[]string{
 				umeeapp.BondDenom,
 			},
 			false,
-			&types.QueryLendAPYResponse{},
+			&types.QuerySupplyAPYResponse{},
 			// Borrow rate * (1 - ReserveFactor - OracleRewardFactor)
 			// 1.50 * (1 - 0.10 - 0.01) = 0.89 * 1.5 = 1.335
-			&types.QueryLendAPYResponse{APY: sdk.MustNewDecFromStr("1.335")},
+			&types.QuerySupplyAPYResponse{APY: sdk.MustNewDecFromStr("1.335")},
 		},
 		testQuery{
 			"query borrow APY",
@@ -329,9 +329,9 @@ func (s *IntegrationTestSuite) TestLeverageScenario() {
 		},
 	}
 
-	lend := testTransaction{
-		"lend",
-		cli.GetCmdLendAsset(),
+	supply := testTransaction{
+		"supply",
+		cli.GetCmdSupply(),
 		[]string{
 			val.Address.String(),
 			"1000uumee",
@@ -443,30 +443,30 @@ func (s *IntegrationTestSuite) TestLeverageScenario() {
 			&types.QueryMarketSizeResponse{MarketSizeUsd: sdk.MustNewDecFromStr("0.03424421")},
 		},
 		testQuery{
-			"query loaned - all",
-			cli.GetCmdQueryLoaned(),
+			"query supplied - all",
+			cli.GetCmdQuerySupplied(),
 			[]string{
 				val.Address.String(),
 			},
 			false,
-			&types.QueryLoanedResponse{},
-			&types.QueryLoanedResponse{
-				Loaned: sdk.NewCoins(
+			&types.QuerySuppliedResponse{},
+			&types.QuerySuppliedResponse{
+				Supplied: sdk.NewCoins(
 					sdk.NewInt64Coin(umeeapp.BondDenom, 1001),
 				),
 			},
 		},
 		testQuery{
-			"query loaned - denom",
-			cli.GetCmdQueryLoaned(),
+			"query supplied - denom",
+			cli.GetCmdQuerySupplied(),
 			[]string{
 				val.Address.String(),
 				fmt.Sprintf("--%s=uumee", cli.FlagDenom),
 			},
 			false,
-			&types.QueryLoanedResponse{},
-			&types.QueryLoanedResponse{
-				Loaned: sdk.NewCoins(
+			&types.QuerySuppliedResponse{},
+			&types.QuerySuppliedResponse{
+				Supplied: sdk.NewCoins(
 					sdk.NewInt64Coin(umeeapp.BondDenom, 1001),
 				),
 			},
@@ -501,6 +501,18 @@ func (s *IntegrationTestSuite) TestLeverageScenario() {
 			},
 		},
 		testQuery{
+			"query total borrowed - denom",
+			cli.GetCmdQueryTotalBorrowed(),
+			[]string{
+				"uumee",
+			},
+			false,
+			&types.QueryTotalBorrowedResponse{},
+			&types.QueryTotalBorrowedResponse{
+				Amount: sdk.NewInt(47),
+			},
+		},
+		testQuery{
 			"query collateral - all",
 			cli.GetCmdQueryCollateral(),
 			[]string{
@@ -530,33 +542,45 @@ func (s *IntegrationTestSuite) TestLeverageScenario() {
 			},
 		},
 		testQuery{
-			"query loaned value - all",
-			cli.GetCmdQueryLoanedValue(),
+			"query total collateral - denom",
+			cli.GetCmdQueryTotalCollateral(),
+			[]string{
+				"u/uumee",
+			},
+			false,
+			&types.QueryTotalCollateralResponse{},
+			&types.QueryTotalCollateralResponse{
+				Amount: sdk.NewInt(1000),
+			},
+		},
+		testQuery{
+			"query supplied value - all",
+			cli.GetCmdQuerySuppliedValue(),
 			[]string{
 				val.Address.String(),
 			},
 			false,
-			&types.QueryLoanedValueResponse{},
-			&types.QueryLoanedValueResponse{
+			&types.QuerySuppliedValueResponse{},
+			&types.QuerySuppliedValueResponse{
 				// From app/test_helpers.go/IntegrationTestNetworkConfig
 				// This result is umee's oracle exchange rate times the
-				// amount loaned.
-				LoanedValue: sdk.MustNewDecFromStr("0.03424421"),
+				// amount supplied.
+				SuppliedValue: sdk.MustNewDecFromStr("0.03424421"),
 				// (1001 / 1000000) umee * 34.21 = 0.03424421
 			},
 		},
 		testQuery{
-			"query loaned value - denom",
-			cli.GetCmdQueryLoanedValue(),
+			"query supplied value - denom",
+			cli.GetCmdQuerySuppliedValue(),
 			[]string{
 				val.Address.String(),
 				fmt.Sprintf("--%s=uumee", cli.FlagDenom),
 			},
 			false,
-			&types.QueryLoanedValueResponse{},
-			&types.QueryLoanedValueResponse{
+			&types.QuerySuppliedValueResponse{},
+			&types.QuerySuppliedValueResponse{
 				// From app/test_helpers.go/IntegrationTestNetworkConfig
-				LoanedValue: sdk.MustNewDecFromStr("0.03424421"),
+				SuppliedValue: sdk.MustNewDecFromStr("0.03424421"),
 				// (1001 / 1000000) umee * 34.21 = 0.03424421
 			},
 		},
@@ -652,7 +676,7 @@ func (s *IntegrationTestSuite) TestLeverageScenario() {
 
 	// These transactions will set up nonzero leverage positions and allow nonzero query results
 	s.runTestCases(
-		lend,
+		supply,
 		addCollateral,
 		borrow,
 		liquidate,
@@ -666,6 +690,6 @@ func (s *IntegrationTestSuite) TestLeverageScenario() {
 		withdraw,
 	)
 
-	// These queries run while the lending and borrowing is active to produce nonzero output
+	// These queries run while the supplying and borrowing is active to produce nonzero output
 	s.runTestCases(nonzeroQueries...)
 }